long long is not supporeted by C90

+++ 0x07-pointers_arrays_strings/8-print_diagsums.c
@@ -9,7 +9,7 @@
 void print_diagsums(int *a, int size)
 {
 	int i;
-	unsigned long long sum1 = 0, sum2 = 0;
+	unsigned long sum1 = 0, sum2 = 0;
 
 	for (i = 0; i < size; i++)
 	{
@@ -17,5 +17,5 @@ void print_diagsums(int *a, int size)
 		sum2 += a[i * size + (size - 1 - i)];
 	}
 
-	printf("%llu, %llu\n", sum1, sum2);
+	printf("%lu, %lu\n", sum1, sum2);
 }
